requirements.txt
django
djangorestframework
psycopg2
drf-yasg

